TransmissionBackend: fix isRunning (openseedbox/openseedbox#105)
Just grepping for a number in the process list in a non-containerized
environment isn't the best way to find out if the backend is running. ;)

$ ps -A | grep 1 | wc -l
361

While checking the backend's availibility through RPC is a good point,
- and the initial commit (d8825f07009b) has it as a hint - it's not
recommended for Transmission, which is retardedly slow on slow hard
disks (due to IO thread and RPC thread being the same)!
See todo.txt at openseedbox/openseedbox@4b12abd0d75 and
transmission/transmission#2462 also!

Fixes: 98fb99538fcd ("mad progress")